Updated author annotations
"author" annotations attached to classes now refer to whoever is most
familiar with the contents of the class, so that any
questions/comments/criticisms can be directed to the right person. It's
not a territory game, I swear.

Also, I compressed doc comments to one line wherever applicable and got
rid of empty comment lines in some places, even though I know that sort
of breaks javadoc formatting. Sorry.